+++ src/Dashboard.js
@@ -1,27 +1,11 @@
 import * as React from 'react';
 import { styled, createTheme, ThemeProvider } from '@mui/material/styles';
-import CssBaseline from '@mui/material/CssBaseline';
-import Box from '@mui/material/Box';
 import MuiAppBar from '@mui/material/AppBar';
-import Toolbar from '@mui/material/Toolbar';
-import List from '@mui/material/List';
-import Typography from '@mui/material/Typography';
-import Divider from '@mui/material/Divider';
-import IconButton from '@mui/material/IconButton';
-import Badge from '@mui/material/Badge';
 import Container from '@mui/material/Container';
-import Grid from '@mui/material/Grid';
-import Paper from '@mui/material/Paper';
-import MenuIcon from '@mui/icons-material/Menu';
-import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
-import NotificationsIcon from '@mui/icons-material/Notifications';
-import { mainListItems } from './components/ListItems';
 import Chart from './components/Chart';
-import { Drawer } from './components/Drawer';
+
 import Network from './components/Network';
-import Orders from './components/Orders';
 import Copyright from './components/Copyright';
-import SanjaalWordCloud from './components/WordCloud';
 const drawerWidth = 240;
 
 const AppBar = styled(MuiAppBar, {
@@ -45,92 +29,22 @@ const AppBar = styled(MuiAppBar, {
 const mdTheme = createTheme();
 
 function DashboardContent() {
-  const [open, setOpen] = React.useState(true);
-  const toggleDrawer = () => {
-    setOpen(!open);
-  };
 
-  return (
-    <ThemeProvider theme={mdTheme}>
-      <Box sx={{ display: 'flex' }}>
-        <CssBaseline />
-        <AppBar position="absolute" open={open}>
-          <Toolbar
-            sx={{
-              pr: '24px', // keep right padding when drawer closed
-            }}
-          >
-            <IconButton
-              edge="start"
-              color="inherit"
-              aria-label="open drawer"
-              onClick={toggleDrawer}
-              sx={{
-                marginRight: '36px',
-                ...(open && { display: 'none' }),
-              }}
-            >
-              <MenuIcon />
-            </IconButton>
-            <Typography
-              component="h1"
-              variant="h6"
-              color="inherit"
-              noWrap
-              sx={{ flexGrow: 1 }}
-            >
-              Twitter Dashboard
-            </Typography>
-          </Toolbar>
-        </AppBar>
-        <Drawer variant="permanent" open={open}>
-          <Toolbar
-            sx={{
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'flex-end',
-              px: [1],
-            }}
-          >
-            <IconButton onClick={toggleDrawer}>
-              <ChevronLeftIcon />
-            </IconButton>
-          </Toolbar>
-          <Divider />
-          <List component="nav">
-            {mainListItems}
-            <Divider sx={{ my: 1 }} />
-          </List>
-        </Drawer>
-        <Box
-          component="main"
-          sx={{
-            backgroundColor: (theme) =>
-              theme.palette.mode === 'light'
-                ? theme.palette.grey[100]
-                : theme.palette.grey[900],
-            flexGrow: 1,
-            height: '100vh',
-            overflow: 'auto',
-          }}
-        >
-          <Toolbar />
-          <Container maxWidth="lg" sx={{ mt: 4, mb: 4 }}>
-          <Chart />
-              <Network/>
-              <SanjaalWordCloud dataUrl={"https://raw.githubusercontent.com/nsuman/SanzalWeb/main/src/user_data.json"} 
-              wordField="name"  weightField="followers_count"/> 
-             
-             <SanjaalWordCloud dataUrl={"https://raw.githubusercontent.com/nsuman/SanzalWeb/main/src/user_data.json"} 
-              wordField="name"  weightField="following_count"/>
-            <Copyright sx={{ pt: 4 }} />
-          </Container>
-        </Box>
-      </Box>
-    </ThemeProvider>
-  );
 }
 
 export default function Dashboard() {
-  return <DashboardContent />;
+  return (
+    <Container maxWidth="lg" sx={{ mt: 4, mb: 4 }}>
+              <Chart />
+              <Network />
+              <WordCloud data_url={"https://raw.githubusercontent.com/nsuman/SanzalWeb/main/src/user_data.json"}
+                wordField="name" weightField="followers_count" />
+
+              <WordCloud data_url={"https://raw.githubusercontent.com/nsuman/SanzalWeb/main/src/user_data.json"}
+                wordField="name" weightField="following_count" />
+
+
+              <Copyright sx={{ pt: 4 }} />
+            </Container>
+  )
 };
